+++ src/test/run-pass/i32-sub.rs
@@ -1,8 +1,8 @@
 // -*- rust -*-
 
 fn main() {
-    let i32 x = i32(-400);
-    x = i32(0) - x;
-    check(x == i32(400));
+    let i32 x = -400_i32;
+    x = 0_i32 - x;
+    check(x == 400_i32);
 }
